Deleted post woring correctly. Added new custom hook for Timeline
